+++ src/Resource/Temporal/Trait/TemporalHelpersTrait.php
@@ -2,11 +2,13 @@
 
 namespace Solital\Core\Resource\Temporal\Trait;
 
+use Solital\Core\Resource\Str\Str;
+
 trait TemporalHelpersTrait
 {
     public static function formatDate(string $date): string
     {
-        if (str_contains($date, "/")) {
+        if (Str::contains($date, "/")) {
             $date = str_replace("/", "-", $date);
         }