+++ src/Security/Exception/HashException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Solital\Core\Security\Exception;
+
+class HashException extends \Exception
+{
+}