+++ tests/Security/HashTest.php
@@ -1,15 +1,24 @@
 <?php
 
 use PHPUnit\Framework\TestCase;
+use Solital\Core\Kernel\Application;
 use Solital\Core\Security\Hash;
 
 require_once dirname(__DIR__) . '/bootstrap.php';
 
 class HashTest extends TestCase
 {
+    public function setUp(): void
+    {
+        Application::getInstance();
+    }
+
     public function testCrypt()
     {
         $hash = Hash::encrypt('word_to_encrypt');
+        $this->assertIsString($hash);
+
+        $hash = encrypt('word_to_encrypt');
         $this->assertIsString($hash);
     }
 
@@ -18,32 +27,31 @@ class HashTest extends TestCase
         $hash = Hash::encrypt('word_to_encrypt');
         $res = Hash::decrypt($hash)->value();
         $this->assertEquals('word_to_encrypt', $res);
+
+        $hash = encrypt('word_to_encrypt');
+        $res = decrypt($hash)->value();
+        $this->assertEquals('word_to_encrypt', $res);
     }
 
-    public function testDecryptIsVaid()
+    public function testDecryptIsValid()
     {
         $hash = Hash::encrypt('word_to_encrypt');
         $res = Hash::decrypt($hash)->isValid();
         $this->assertTrue($res);
+
+        $hash = encrypt('word_to_encrypt');
+        $res = decrypt($hash)->isValid();
+        $this->assertTrue($res);
     }
 
-    public function testSodium()
+    public function testDecryptIsNotValid()
     {
-        $isset = Hash::checkSodium();
-        $this->assertTrue($isset);
-    }
+        $hash = Hash::encrypt('word_to_encrypt', '-1 hour');
+        $res = Hash::decrypt($hash)->isValid();
+        $this->assertFalse($res);
 
-    public function testSodiumCrypt()
-    {
-        $encoded = Hash::sodiumCrypt("word_to_encrypt");
-        $this->assertIsString($encoded);
-    }
-
-    public function testSodiumDecrypt()
-    {
-        $key = Hash::getSodiumKey();
-        $encoded = Hash::sodiumCrypt("word_to_encrypt", $key);
-        $decoded = Hash::sodiumDecrypt($encoded, $key);
-        $this->assertIsString($decoded);
+        $hash = encrypt('word_to_encrypt', '-1 hour');
+        $res = decrypt($hash)->isValid();
+        $this->assertFalse($res);
     }
 }