+++ tests/Wolf/WolfTest.php
@@ -50,7 +50,7 @@ class WolfTest extends TestCase
         $this->assertEquals('<h1>home test SOLITAL</h1><p>solital@email.com</p>', $res);
     }
 
-    /* public function testCacheView()
+    public function testCacheView()
     {
         $result = false;
 
@@ -61,7 +61,7 @@ class WolfTest extends TestCase
         ]);
 
         $wolf->setView('contact')->setCacheTime('week')->render();
-    } */
+    }
 
     public function testWithInstance()
     {